Refine the implementation
